refactor(rfc64): split SWM inventory persistence seams
