+++ packages/agent/scripts/test-package-root.mjs
@@ -124,8 +124,11 @@ const blockedRfc64Modules = [
   'control-object-store-v1.js',
   'durable-file-store-v1.js',
   'inventory-v1/exact-record.js',
+  'inventory-v1/swm-author-inventory-commit-plan.js',
   'inventory-v1/swm-author-inventory-contracts.js',
+  'inventory-v1/swm-author-inventory-mutation.js',
   'inventory-v1/swm-author-inventory-persistence.js',
+  'inventory-v1/swm-author-inventory-sql-codec.js',
   'finalized-vm-agent-precommit-v1.js',
   'finalized-vm-composer-v1.js',
   'finalized-vm-runtime-v1.js',
